requirements_dev.txt
coverage==4.4.1
dataset==0.8.0
Faker==0.8.1
numpy==1.13.1
pandas==0.20.3
pylint==1.7.2
pytest-cov==2.5.1
Sphinx==1.6.3
sphinx-autobuild==0.7.1
tox==2.8.1
